+++ R/build_gho.R
@@ -1,30 +1,19 @@
 #' Check and Build a gho Object
 #'
-#' @param x A vector of codes.
-#' @param labels A vector of labels.
-#' @param attrs A \code{data_frame} of attributes.
+#' @param x A ODataQuery object
 #'
 #' @return A \code{gho} object.
 #' @keywords internal
-build_gho <- function(x, labels, attrs = NULL) {
-  return_if_message(x)
-  stopifnot(
-    is.null(dim(x)),
-    is.null(dim(labels)),
-    is.character(x),
-    is.character(labels),
-    is.data.frame(attrs) | is.null(attrs),
-    length(x) == length(labels),
-    if (! is.null(attrs)) nrow(attrs) == length(x) else TRUE
-  )
+#'
 
-  if (! is.null(attrs)) {
-    if (ncol(attrs) <= 1) attrs <- NULL
-  }
-  structure(
-    as.vector(x),
-    labels = as.vector(labels),
-    attrs = attrs,
-    class = "gho"
-  )
+build_gho <- function(x){
+  w <- graceful_fail(x$url, config = c(
+    httr::user_agent("https://pierucci.org/rgho/")
+  ))
+  return_if_message(w, display = TRUE, n = 2)
+  ret <- x$retrieve()
+  structure(ret$value,
+            class = c("gho", "data.frame"),
+            url = ret$`@odata.context`)
 }
+