+++ R/get_object.R
@@ -1,24 +1,8 @@
 #' @rdname man-gho
 #' @export
 get_gho_dimensions <- function() {
-  resp <- get_gho(
-    url = build_gho_url(dimension = NULL, format = "json")
-  )
-  return_if_message(resp)
-
-  if (httr::http_type(resp) != "application/json") {
-    stop(sprintf(
-      "Unexpected type: '%s', expected 'application/json'.",
-      httr::http_type(resp)
-    ))
-  }
-
-  tab_res <- parse_json(resp, type = "dimension")
-
-  build_gho(
-    tab_res$label, labels = tab_res$display,
-    attrs = dplyr::select(tab_res, -.data$label, -.data$display)
-  )
+  resp <- get_gho()$path("Dimension")
+  build_gho(resp)
 }
 
 #' @rdname man-gho
@@ -26,77 +10,13 @@ get_gho_dimensions <- function() {
 get_gho_codes <- function(dimension = "GHO") {
   dims <- get_gho_dimensions()
   return_if_message(dims)
-  if(!length(dims) & !is.null(attr(dims, "message"))) return(invisible(dims))
   stopifnot(
-    dimension %in% get_gho_dimensions()
+    dimension %in% dims$Code
   )
 
-  resp <- get_gho(
-    url = build_gho_url(dimension = dimension, format = "json")
-  )
+  vals <- get_gho()$path(sprintf("DIMENSION/%s/DimensionValues", dimension))
 
-  if(!length(resp) & !is.null(attr(resp, "message"))) return(invisible(resp))
-
-  if (httr::http_type(resp) != "application/json") {
-    stop(sprintf(
-      "Unexpected type: '%s', expected 'application/json'.",
-      httr::http_type(resp)
-    ))
-  }
-
-  tab_res <- parse_json(resp, type = "code")
-
-  build_gho(
-    tab_res$label, labels = tab_res$display,
-    attrs = dplyr::select(tab_res, -.data$label, -.data$display)
-  )
-}
-
-parse_json <- function(resp, type = c("dimension", "code")) {
-  if (is.null(resp)) return(NULL)
-  type <- match.arg(type)
-
-  list_resp <- resp %>%
-    httr::content(as = "text") %>%
-    clean_json %>%
-    jsonlite::fromJSON(
-      simplifyVector = F,
-      simplifyDataFrame = F,
-      simplifyMatrix = F
-    )
-
-  switch(
-    type,
-    dimension = {
-      x <- list_resp$dimension
-    },
-    code = {
-      x <- list_resp$dimension[[1]]$code
-    }
-  )
-
-  g <- function(x, y) rbind(unlist(x), unlist(y))
-
-  f <- function(x) {
-    if (length(x$attr) > 1) {
-      mat_attr <- Reduce(g, x$attr)
-    } else {
-      mat_attr <- rbind(unlist(x$attr))
-    }
-
-    res <- rbind(
-      matrix(
-        c("label", "display", x$label, x$display),
-        nrow = 2
-      ),
-      mat_attr
-    )
-
-    cbind(res, x$label)
-  }
-
-  do.call(rbind, lapply(x, f)) %>%
-    tibble::as_tibble(.name_repair = "unique") %>%
-    stats::setNames(c("category", "value", "code")) %>%
-    tidyr::spread_(key_col = "category", value_col = "value")
+  res <- build_gho(vals)
+  structure(res[c("Code", "Title")],
+            url = attr(res, "url"))
 }