+++ R/zzz.R
@@ -1,12 +1,9 @@
 .onLoad <- function(libname, pkgname) {
   op <- options()
   op.rgho <- list(
-    rgho.verbose = FALSE,
-    rgho.retry = 5,
-    rgho.memotime = 3600,
     rgho.http_proxy = NULL,
     rgho.n = 6,
-    rgho.baseurl = "https://apps.who.int/gho/athena/api/"
+    rgho.baseurl = "https://ghoapi.azureedge.net/api/"
   )
   toset <- !(names(op.rgho) %in% names(op))
   if(any(toset)) options(op.rgho[toset])