+++ vignettes/c-codes-gho.R
@@ -1,4 +1,0 @@
-## ---- echo=FALSE, include=FALSE-----------------------------------------------
-library(rgho)
-library(knitr)
-